+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The new `supernotify.notify` action introduced in v2.0.0 has a simpler way of ha
 - `message_html` removed from spoken only envelopes
 ### Media Transport
 - Now uses the standard image grabbing modules, and supports the `jpeg_opts` and `png_opts` for image tuning
+- Added missing `snapshot_image_path` to allowed `media:` options
 ### Documentation
 - Fix automatically generated validation schema documentation
 - Added automated test for the YAML examples in docs